pyproject.toml
[tool.poetry]
name = "notiflow"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
supabase = "^2.15.0"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"
databases = "^0.9.0"
requests = "^2.32.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
